pyproject.toml
[project]
name = "a-star"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "imageio==2.37.2",
    "mouseinfo==0.1.3",
    "numpy==2.3.4",
    "pillow==12.0.0",
    "pyautogui==0.9.54",
    "pygame-ce==2.5.6",
    "pygetwindow==0.0.9",
    "pymsgbox==2.0.1",
    "pyperclip==1.11.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "pytweening==1.2.0",
    "setuptools==80.9.0",
]
